requirements.txt
streamlit
torch
torchvision
numpy
Pillow
joblib
scikit-learn
